Refactor the freeradius-eduroam module to make it more generic
Part-way through upgrading to 3.x, we realised that the PAM-specific
component of this role was miniscule, and what we really had was a nice
generic role that configured an eduroam IdP. On that realisation, and
after brief correspondence with @chrohrer, we've refactored this role
to remove explicit mention of -pam and make it a configuration option.

Some minor improvements have been made along the way, such as making the
radius_test_user variable substantially more generic.