Add (disabled) listen stanzas to eduroam virtual server
The default block contains listen stanzas. If people have deleted that,
it means the server does not listen on any useful ports. However, if it
is still in place, it conflicts with listen stanzas in any other virtual
server. Thus we document how people can do this themselves if they need to.